refactor: refactor main data flow
(cherry picked from commit 1c2096d367c0e5a0842f4236ef7e7152e70564ad)
